requirements.txt
asgiref==3.6.0
distlib==0.3.6
Django==4.2.1
django-cors-headers==4.0.0
djangorestframework==3.14.0
filelock==3.11.0
platformdirs==3.2.0
PuLP==2.7.0
pytz==2023.3
six==1.16.0
sqlparse==0.4.4
virtualenv==20.21.0
transformers==4.25.1
bentoml==0.13.1
torch